Test fixes, resolve TODO, and small refactoring
