+++ Source/DafnyCore/AST/Modules/TopLevelDeclarations.cs
