+++ Test/git-issues/git-issue-3451/git-issue-3451a.dfy.expect
@@ -1,3 +1,3 @@
-git-issue-3451a.dfy(1,0): Warning: File contains no code
 git-issue-3451a.dfy(5,0): Error: this symbol not expected in Dafny
+git-issue-3451a.dfy(1,0): Warning: File contains no code
 1 parse errors detected in git-issue-3451a.dfy
