+++ Test/separate-verification/assumptions.dfy.expect
@@ -9,6 +9,7 @@ TestAuditor.dfy(126,2): Warning: /!\ No terms found to trigger on.
 
 Dafny program verifier did not attempt verification
 assumptions.dfy(6,8): Error: the included file TestAuditor.dfy contains error(s)
+TestAuditor.dfy(89,27): Error: Trait method calls may not terminate (uses [{:termination false}]).
 TestAuditor.dfy(22,6): Error: Ghost declaration has no body and has an ensures clause.
 TestAuditor.dfy(27,15): Error: Ghost declaration has no body and has an ensures clause.
 TestAuditor.dfy(45,17): Error: Declaration with [{:extern}] has a ensures clause.
@@ -29,5 +30,4 @@ TestAuditor.dfy(130,2): Error: Definition contains loop with no body.
 TestAuditor.dfy(137,2): Error: Assertion has explicit temporary [{:only}] attribute.
 TestAuditor.dfy(11,8): Error: the included file IgnoredAssumptions.dfy contains error(s)
 IgnoredAssumptions.dfy(2,15): Error: Ghost declaration has no body and has an ensures clause.
-TestAuditor.dfy(89,27): Error: Trait method calls may not terminate (uses [{:termination false}]).
 Wrote textual form of partial target program to assumptions-lib/assumptions.doo
